+++ WordPress-Restore.sh
@@ -8,6 +8,12 @@ source /Joy/Utility/Script/Common.sh
 
 # Write your code below ---
 
+# Check 1: Prevent root execution
+if [[ $IS_ROOT_USER -eq 1 ]]; then
+    echo "ERROR: This script should NOT run as root!" >&2
+    exit 1
+fi
+
 BFN=$1
 unzip -o $BFN.zip` && echo "Press ENTER to edit 'wp-config.php' to set new database credentials..." && read DUMMY_INPUT && nano wp-config.php && wp db import wordpress.sql && rm -f wordpress.sql
